+++ apps/zui/src/js/state/Tabs/selectors.ts
@@ -3,11 +3,6 @@ import {State} from "../types"
 import {createIsEqualSelector} from "../utils"
 import {findQuerySessionTab} from "./find"
 import {activeTabsSelect, getActiveTabs} from "../Window/selectors"
-import tab from "src/js/models/tab"
-import Pools from "../Pools"
-import Lakes from "../Lakes"
-import Current from "../Current"
-import Queries from "../Queries"
 
 export const getData = activeTabsSelect((tabs) => tabs.data)
 export const getActive = activeTabsSelect((tabs) => tabs.active)
@@ -37,14 +32,3 @@ export const findFirstQuerySession = createSelector(
 
 export const findById = (tabId: string) =>
   createSelector(getData, (tabs) => tabs.find((t) => t.id === tabId))
-
-export const getTabModels = createSelector(
-  getIds,
-  Pools.raw,
-  Lakes.raw,
-  Current.getLakeId,
-  Queries.getQueryIdToName,
-  (ids, pools, lakes, lakeId, queryIdToName) => {
-    return ids.map((id) => tab(id, lakes, pools, queryIdToName, lakeId))
-  }
-)
